package structure refactored
