+++ src/main/kotlin/io/github/janbarari/gradle/ExcludeJacocoGenerated.kt
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.github.janbarari.gradle.extension
+package io.github.janbarari.gradle
 
 /**
  * In order to have reliable code coverage sometimes it's needed to exclude some
