+++ src/main/kotlin/io/github/janbarari/gradle/analytics/GradleAnalyticsPlugin.kt
@@ -24,7 +24,7 @@ package io.github.janbarari.gradle.analytics
 
 import io.github.janbarari.gradle.analytics.reporttask.ReportAnalyticsTask
 import io.github.janbarari.gradle.analytics.scanner.ScannerUtils
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 import io.github.janbarari.gradle.utils.ProjectUtils
 import org.gradle.api.Plugin
 import org.gradle.api.Project
