+++ src/main/kotlin/io/github/janbarari/gradle/analytics/GradleAnalyticsPluginConfig.kt
@@ -25,7 +25,7 @@ package io.github.janbarari.gradle.analytics
 import io.github.janbarari.gradle.analytics.data.database.connection.DatabaseConnection
 import io.github.janbarari.gradle.analytics.data.database.connection.MySqlDatabaseConnection
 import io.github.janbarari.gradle.analytics.data.database.connection.SqliteDatabaseConnection
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 import org.gradle.api.Project
 
 /**
