+++ src/main/kotlin/io/github/janbarari/gradle/analytics/data/database/Database.kt
@@ -28,7 +28,7 @@ import io.github.janbarari.gradle.analytics.data.database.connection.MySqlDataba
 import io.github.janbarari.gradle.analytics.data.database.connection.SqliteDatabaseConnection
 import io.github.janbarari.gradle.analytics.data.database.table.MetricTable
 import io.github.janbarari.gradle.analytics.data.database.table.TemporaryMetricTable
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 import io.github.janbarari.gradle.extension.isNotNull
 import io.github.janbarari.gradle.extension.isNull
 import io.github.janbarari.gradle.extension.toRealPath
