+++ src/main/kotlin/io/github/janbarari/gradle/analytics/data/database/connection/SqliteDatabaseConnection.kt
@@ -23,7 +23,7 @@
 package io.github.janbarari.gradle.analytics.data.database.connection
 
 import io.github.janbarari.gradle.analytics.GradleAnalyticsPlugin.Companion.PLUGIN_NAME
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 
 /**
  * @author Mehdi-Janbarari
