+++ src/main/kotlin/io/github/janbarari/gradle/analytics/domain/model/BuildMetric.kt
@@ -1,7 +1,7 @@
 package io.github.janbarari.gradle.analytics.domain.model
 
 import com.squareup.moshi.Json
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 
 @ExcludeJacocoGenerated
 data class BuildMetric(
