+++ src/main/kotlin/io/github/janbarari/gradle/analytics/domain/model/HardwareInfo.kt
@@ -22,7 +22,7 @@
  */
 package io.github.janbarari.gradle.analytics.domain.model
 
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 
 /**
  * @author Mehdi-Janbarari
