+++ src/main/kotlin/io/github/janbarari/gradle/analytics/reporttask/exception/MissingPropertyException.kt
@@ -22,7 +22,7 @@
  */
 package io.github.janbarari.gradle.analytics.reporttask.exception
 
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 
 /**
  * @author Mehdi-Janbarari
