+++ src/main/kotlin/io/github/janbarari/gradle/analytics/scanner/ScannerUtils.kt
@@ -27,7 +27,7 @@ import io.github.janbarari.gradle.analytics.scanner.configuration.BuildConfigura
 import io.github.janbarari.gradle.analytics.scanner.dependencyresolution.BuildDependencyResolutionService
 import io.github.janbarari.gradle.analytics.scanner.execution.BuildExecutionService
 import io.github.janbarari.gradle.analytics.scanner.initialization.BuildInitializationService
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 import io.github.janbarari.gradle.extension.envCI
 import io.github.janbarari.gradle.extension.getRequestedTasks
 import org.gradle.api.Project
