+++ src/main/kotlin/io/github/janbarari/gradle/analytics/scanner/initialization/BuildInitializationService.kt
@@ -22,7 +22,7 @@
  */
 package io.github.janbarari.gradle.analytics.scanner.initialization
 
-import io.github.janbarari.gradle.extension.ExcludeJacocoGenerated
+import io.github.janbarari.gradle.ExcludeJacocoGenerated
 import org.gradle.BuildResult
 import org.gradle.api.Project
 import org.gradle.api.ProjectEvaluationListener
